backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
firebase-admin==6.2.0
python-decouple==3.8
Pillow==10.0.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0
openpyxl==3.1.2
reportlab==4.0.7
psycopg2-binary==2.9.10
